pyproject.toml
[project]
name = "retail-monitor"
version = "0.1.0"
description = "Single-account stock monitor with browser session handoff"
requires-python = ">=3.11"
dependencies = [
    "playwright>=1.52.0",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
]

[tool.setuptools]
py-modules = []
